flash c6 chip with 4 separate .bin images

+++ main/get_firmware.sh
@@ -9,23 +9,31 @@ if [ -z "$VERSION" ];  then
 fi
 TAG=v$VERSION
 
-ARTIFACT_FILENAME=firmware-why2025-carrier-$VERSION.bin
-FILENAME=firmware-why2025-carrier.bin
-#mkdir -p build
+RELEASE_JSON=$(curl -L \
+                 -H "Accept: application/vnd.github+json" \
+                 -H "X-GitHub-Api-Version: 2022-11-28" \
+                 https://api.github.com/repos/$OWNER/$REPO/releases/tags/$TAG \
+                 | tee /dev/stderr)
 
-ARTIFACT_JSON=$(curl -L \
-  -H "Accept: application/vnd.github+json" \
-  -H "X-GitHub-Api-Version: 2022-11-28" \
-  https://api.github.com/repos/$OWNER/$REPO/releases/tags/$TAG \
-  | tee /dev/stderr | jq ".assets[] | select(.name==\""${ARTIFACT_FILENAME}"\")")
+for ARTIFACT_FILENAME in boot_app0.bin bootloader.bin firmware.bin partitions.bin ; do
+  FILENAME=c6-$ARTIFACT_FILENAME
 
-DIGEST=$(echo "$ARTIFACT_JSON" | jq -r .digest)
-DOWNLOAD_URL=$(echo "$ARTIFACT_JSON" | jq -r .browser_download_url)
+  ARTIFACT_JSON=$(echo $RELEASE_JSON | jq ".assets[] | select(.name==\""${ARTIFACT_FILENAME}"\")")
 
-curl -L -o $FILENAME $DOWNLOAD_URL
-CHECKSUM=$(echo $DIGEST | cut -d: -f2 | tee $FILENAME.checksum)
-ALGO=$(echo $DIGEST | cut -d: -f1 | sed "s/sha//")
+  DIGEST=$(echo "$ARTIFACT_JSON" | jq -r .digest)
+  DOWNLOAD_URL=$(echo "$ARTIFACT_JSON" | jq -r .browser_download_url)
 
-shasum --alg $ALGO $FILENAME | grep $CHECKSUM
+  if [ ! -f $FILENAME ]; then
+    curl -L -o $FILENAME $DOWNLOAD_URL
+  fi
 
-md5sum $FILENAME | cut -d' ' -f1 > $FILENAME.md5
+  CHECKSUM=$(echo $DIGEST | cut -d: -f2 | tee $FILENAME.checksum)
+  ALGO=$(echo $DIGEST | cut -d: -f1 | sed "s/sha//")
+
+  if ! (shasum --alg "$ALGO" "$FILENAME" | grep "$CHECKSUM"); then
+    echo "$FILENAME did not match expected checksum $CHECKSUM!"
+    exit 1
+  fi
+
+  md5sum $FILENAME | cut -d' ' -f1 > $FILENAME.md5
+done
